+++ src/api_bak/ing/comment.rs
@@ -1,4 +1,4 @@
-use crate::api::ing::Ing;
+use crate::api_bak::ing::Ing;
 use crate::infra::http::{RequestBuilderExt, unit_or_err};
 use crate::openapi;
 use anyhow::Result;
