+++ src/api_bak/ing/get_comment_list.rs
@@ -1,4 +1,4 @@
-use crate::api::ing::Ing;
+use crate::api_bak::ing::Ing;
 use crate::infra::http::{RequestBuilderExt, body_or_err};
 use crate::infra::json;
 use crate::infra::result::WrapResult;
