+++ src/api_bak/mod.rs
@@ -9,7 +9,7 @@ pub const BLOG_BACKEND: &str = "https://i.cnblogs.com/api";
 #[macro_export]
 macro_rules! blog_backend {
     ($($arg:tt)*) => {{
-        use $crate::api::BLOG_BACKEND;
+        use $crate::api_bak::BLOG_BACKEND;
         format!("{}{}", BLOG_BACKEND, format_args!($($arg)*))
     }};
 }
@@ -18,7 +18,7 @@ pub const OPENAPI: &str = "https://api.cnblogs.com/api";
 #[macro_export]
 macro_rules! openapi {
     ($($arg:tt)*) => {{
-        use $crate::api::OPENAPI;
+        use $crate::api_bak::OPENAPI;
         format!("{}{}", OPENAPI, format_args!($($arg)*))
     }};
 }
@@ -27,7 +27,7 @@ pub const OAUTH: &str = "https://oauth.cnblogs.com";
 #[macro_export]
 macro_rules! oauth {
     ($($arg:tt)*) => {{
-        use $crate::api::OAUTH;
+        use $crate::api_bak::OAUTH;
         format!("{}{}", OAUTH, format_args!($($arg)*))
     }};
 }
