+++ src/api_bak/post/get_comment_list.rs
@@ -1,5 +1,5 @@
-use crate::api::post::Post;
-use crate::api::user::User;
+use crate::api_bak::post::Post;
+use crate::api_bak::user::User;
 use crate::infra::http::{RequestBuilderExt, body_or_err};
 use crate::infra::json;
 use crate::infra::result::WrapResult;
