+++ src/api_bak/user/info.rs
@@ -1,4 +1,4 @@
-use crate::api::user::User;
+use crate::api_bak::user::User;
 use crate::infra::http::{RequestBuilderExt, body_or_err};
 use crate::infra::json;
 use crate::infra::result::WrapResult;
