+++ src/apis/ing/mod.rs
@@ -20,7 +20,7 @@ use reqwest::{Client, Response};
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    api::ing::{IngSendFrom, get_list::IngEntry},
+    api_bak::ing::{IngSendFrom, get_list::IngEntry},
     infra::http::RequestBuilderExt,
     openapi,
 };
