+++ src/bin/cnb.rs
@@ -9,12 +9,12 @@ extern crate cnblogs_lib;
 use anyhow::Result;
 use clap::Parser;
 use clap::{Command, CommandFactory};
-use cnblogs_lib::api::auth::session;
-use cnblogs_lib::api::fav::Fav;
-use cnblogs_lib::api::ing::Ing;
-use cnblogs_lib::api::news::News;
-use cnblogs_lib::api::post::Post;
-use cnblogs_lib::api::user::User;
+use cnblogs_lib::api_bak::auth::session;
+use cnblogs_lib::api_bak::fav::Fav;
+use cnblogs_lib::api_bak::ing::Ing;
+use cnblogs_lib::api_bak::news::News;
+use cnblogs_lib::api_bak::post::Post;
+use cnblogs_lib::api_bak::user::User;
 use cnblogs_lib::args::cmd::post::{CreateCmd, UpdateCmd};
 use cnblogs_lib::args::parser::no_operation;
 use cnblogs_lib::args::{Args, parser};
