+++ src/display/colorful/fav.rs
@@ -1,4 +1,4 @@
-use crate::api::fav::get_list::FavEntry;
+use crate::api_bak::fav::get_list::FavEntry;
 use crate::args::TimeStyle;
 use crate::display::colorful::fmt_err;
 use crate::infra::result::WrapResult;
