+++ src/display/colorful/news.rs
@@ -1,4 +1,4 @@
-use crate::api::news::get_list::NewsEntry;
+use crate::api_bak::news::get_list::NewsEntry;
 use crate::args::TimeStyle;
 use crate::display::colorful::fmt_err;
 use crate::infra::result::WrapResult;
