+++ src/display/colorful/user.rs
@@ -1,4 +1,4 @@
-use crate::api::user::info::UserInfo;
+use crate::api_bak::user::info::UserInfo;
 use crate::display::colorful::fmt_err;
 use crate::infra::result::WrapResult;
 use anyhow::Result;
