+++ src/display/json/ing.rs
@@ -1,5 +1,5 @@
-use crate::api::ing::get_comment_list::IngCommentEntry;
-use crate::api::ing::get_list::IngEntry;
+use crate::api_bak::ing::get_comment_list::IngCommentEntry;
+use crate::api_bak::ing::get_list::IngEntry;
 use crate::display::json::{fmt_err, fmt_ok};
 use anyhow::Result;
 use serde_json::json;
