+++ src/display/json/user.rs
@@ -1,4 +1,4 @@
-use crate::api::user::info::UserInfo;
+use crate::api_bak::user::info::UserInfo;
 use crate::display::json::fmt_result;
 use anyhow::Result;
 use serde_json::json;
