+++ src/display/mod.rs
@@ -1,11 +1,11 @@
-use crate::api::fav::get_list::FavEntry;
-use crate::api::ing::get_comment_list::IngCommentEntry;
-use crate::api::ing::get_list::IngEntry;
-use crate::api::news::get_list::NewsEntry;
-use crate::api::post::get_comment_list::PostCommentEntry;
-use crate::api::post::get_one::PostEntry;
-use crate::api::post::search_site::SearchResultEntry;
-use crate::api::user::info::UserInfo;
+use crate::api_bak::fav::get_list::FavEntry;
+use crate::api_bak::ing::get_comment_list::IngCommentEntry;
+use crate::api_bak::ing::get_list::IngEntry;
+use crate::api_bak::news::get_list::NewsEntry;
+use crate::api_bak::post::get_comment_list::PostCommentEntry;
+use crate::api_bak::post::get_one::PostEntry;
+use crate::api_bak::post::search_site::SearchResultEntry;
+use crate::api_bak::user::info::UserInfo;
 use crate::args::{Style, TimeStyle};
 use crate::infra::result::WrapResult;
 use anyhow::Result;
