+++ src/display/normal/post.rs
@@ -1,6 +1,6 @@
-use crate::api::post::get_comment_list::PostCommentEntry;
-use crate::api::post::get_one::PostEntry;
-use crate::api::post::search_site::SearchResultEntry;
+use crate::api_bak::post::get_comment_list::PostCommentEntry;
+use crate::api_bak::post::get_one::PostEntry;
+use crate::api_bak::post::search_site::SearchResultEntry;
 use crate::args::TimeStyle;
 use crate::display::normal::fmt_err;
 use crate::infra::result::WrapResult;
