+++ src/display/normal/user.rs
@@ -1,4 +1,4 @@
-use crate::api::user::info::UserInfo;
+use crate::api_bak::user::info::UserInfo;
 use crate::display::normal::fmt_err;
 use crate::infra::result::WrapResult;
 use anyhow::Result;
