+++ src/lib.rs
@@ -4,7 +4,7 @@
 #![feature(iterator_try_reduce)]
 #![warn(clippy::all, clippy::nursery, clippy::cargo_common_metadata)]
 
-pub mod api;
+pub mod api_bak;
 pub mod apis;
 pub mod args;
 pub mod display;
